refactor: split place detail sheet responsibilities
