+++ src/components/PlaceDetailSheet.tsx
@@ -1,36 +1,10 @@
-﻿import { useRef } from 'react';
 import { categoryInfo } from '../lib/categories';
 import { PlaceBadgeRow } from './place/PlaceBadgeRow';
 import { PlaceDetailHeader } from './place/PlaceDetailHeader';
+import { PlaceDetailReviewSection } from './place/PlaceDetailReviewSection';
 import { PlaceProofCard } from './place/PlaceProofCard';
-import { PlaceReviewPreviewList } from './review/PlaceReviewPreviewList';
-import { ReviewComposer } from './ReviewComposer';
-import type { ApiStatus, DrawerState, Place, Review, ReviewMood, StampLog } from '../types';
-
-interface PlaceDetailSheetProps {
-  place: Place | null;
-  reviews: Review[];
-  isOpen: boolean;
-  drawerState: DrawerState;
-  loggedIn: boolean;
-  visitCount: number;
-  latestStamp: StampLog | null;
-  todayStamp: StampLog | null;
-  hasCreatedReviewToday: boolean;
-  stampActionStatus: ApiStatus;
-  stampActionMessage: string;
-  reviewProofMessage: string;
-  reviewError: string | null;
-  reviewSubmitting: boolean;
-  canCreateReview: boolean;
-  onOpenFeedReview: () => void;
-  onClose: () => void;
-  onExpand: () => void;
-  onCollapse: () => void;
-  onRequestLogin: () => void;
-  onClaimStamp: (place: Place) => Promise<void>;
-  onCreateReview: (payload: { stampId: string; body: string; mood: ReviewMood; file: File | null }) => Promise<void>;
-}
+import type { PlaceDetailSheetProps } from './place/placeDetailSheetTypes';
+import { usePlaceDrawerHandle } from './place/usePlaceDrawerHandle';
 
 export function PlaceDetailSheet({
   place,
@@ -56,44 +30,21 @@ export function PlaceDetailSheet({
   onClaimStamp,
   onCreateReview,
 }: PlaceDetailSheetProps) {
-  const dragStartYRef = useRef<number | null>(null);
+  const { handlePointerDown, handlePointerUp, handleClick } = usePlaceDrawerHandle({
+    drawerState,
+    onClose,
+    onExpand,
+    onCollapse,
+  });
 
   if (!place || !isOpen) {
     return null;
-  }
-
-  function handlePointerDown(event: React.PointerEvent<HTMLButtonElement>) {
-    dragStartYRef.current = event.clientY;
-  }
-
-  function handlePointerUp(event: React.PointerEvent<HTMLButtonElement>) {
-    if (dragStartYRef.current === null) {
-      return;
-    }
-
-    const delta = event.clientY - dragStartYRef.current;
-    dragStartYRef.current = null;
-
-    if (delta > 72) {
-      if (drawerState === 'full') {
-        onCollapse();
-        return;
-      }
-      onClose();
-      return;
-    }
-
-    if (delta < -48) {
-      onExpand();
-    }
   }
 
   const sheetClassName = `place-drawer place-drawer--${drawerState}`;
   const visitLabel = latestStamp ? latestStamp.visitLabel : '첫 방문 대기';
   const canClaimStamp = loggedIn && !todayStamp;
   const categoryMeta = categoryInfo[place.category];
-  const reviewPreview = reviews.slice(0, 2);
-  const reviewComposerStatus = !loggedIn ? 'login' : hasCreatedReviewToday ? 'daily-limit' : todayStamp ? 'ready' : 'claim';
 
   return (
     <section className={sheetClassName} aria-label="장소 상세 시트">
@@ -103,7 +54,7 @@ export function PlaceDetailSheet({
         aria-label="시트 높이 조절"
         onPointerDown={handlePointerDown}
         onPointerUp={handlePointerUp}
-        onClick={drawerState === 'partial' ? onExpand : onCollapse}
+        onClick={handleClick}
       >
         <span />
       </button>
@@ -143,43 +94,21 @@ export function PlaceDetailSheet({
           <p>{place.routeHint}</p>
         </div>
 
-        <ReviewComposer
-          placeName={place.name}
+        <PlaceDetailReviewSection
+          place={place}
+          reviews={reviews}
           loggedIn={loggedIn}
-          canSubmit={canCreateReview}
-          status={reviewComposerStatus}
-          submitting={reviewSubmitting}
-          errorMessage={reviewError}
-          proofMessage={reviewProofMessage}
-          onSubmit={({ body, mood, file }) => {
-            if (!todayStamp) {
-              return Promise.resolve();
-            }
-            return onCreateReview({ stampId: todayStamp.id, body, mood, file });
-          }}
+          todayStamp={todayStamp}
+          hasCreatedReviewToday={hasCreatedReviewToday}
+          reviewSubmitting={reviewSubmitting}
+          reviewError={reviewError}
+          reviewProofMessage={reviewProofMessage}
+          canCreateReview={canCreateReview}
+          onOpenFeedReview={onOpenFeedReview}
           onRequestLogin={onRequestLogin}
-          onRequestProof={() => {
-            if (!loggedIn) {
-              onRequestLogin();
-              return;
-            }
-            if (!todayStamp) {
-              void onClaimStamp(place);
-            }
-          }}
+          onClaimStamp={onClaimStamp}
+          onCreateReview={onCreateReview}
         />
-
-        <div className="section-title-row section-title-row--tight">
-          <div>
-            <p className="eyebrow">PLACE FEED</p>
-            <h3>이 장소 피드</h3>
-          </div>
-          <button type="button" className="secondary-button place-drawer__feed-button" onClick={onOpenFeedReview}>
-            피드에서 보기
-          </button>
-        </div>
-
-        <PlaceReviewPreviewList reviews={reviewPreview} />
       </div>
     </section>
   );
